+++ tests/unit/model/test_models.py
@@ -19,6 +19,7 @@ import inspect
 
 import pytest
 
+from grand_trade_auto.model import company
 from grand_trade_auto.model import datafeed_src
 from grand_trade_auto.model import exchange
 
@@ -27,6 +28,7 @@ from grand_trade_auto.model import exchange
 @pytest.mark.parametrize('model_cls, extra_attrs', [
     (datafeed_src.DatafeedSrc, []),
     (exchange.Exchange, []),
+    (company.Company, []),
 ])
 def test_model_attrs(model_cls, extra_attrs):
     """